chore: refactor move zod type
