+++ packages/types/VideoApiAdapter.d.ts
@@ -1,10 +1,9 @@
 import type {
-  GetRecordingsResponseSchema,
-  GetAccessLinkResponseSchema,
   TSubmitBatchProcessorJobRes,
   batchProcessorBody,
   TGetTranscriptAccessLink,
-} from "@calcom/prisma/zod-utils";
+} from "@calcom/app-store/dailyvideo/zod";
+import type { GetRecordingsResponseSchema, GetAccessLinkResponseSchema } from "@calcom/prisma/zod-utils";
 
 import type { EventBusyDate } from "./Calendar";
 import type { CredentialPayload } from "./Credential";
